requirements.txt
alembic==1.3.1
aniso8601==8.0.0
attrs==19.3.0
backcall==0.1.0
Click==7.0
decorator==4.4.1
Flask==1.1.1
Flask-Cache==0.13.1
Flask-JWT==0.3.2
flask-marshmallow==0.10.1
Flask-RESTful==0.3.7
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.0.0
ipython==7.10.2
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
logzero==1.5.0
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.21.0
more-itertools==8.0.0
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.2
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
PyJWT==1.4.2
pyrsistent==0.15.6
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-json-logger==0.1.11
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.11
SQLAlchemy-Utils==0.36.0
traitlets==4.3.3
uuid==1.30
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
